requirements.txt
discord.py==2.3.1
python-dotenv~=1.0.0
hashids~=1.3.1
aiosqlite~=0.19.0
discord~=2.3.1
aiofiles~=23.1.0